requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python-decouple==3.8
drf-yasg==1.21.11
psycopg2-binary==2.9.9
django-cors-headers==4.9.0